outline-generator/requirements.txt
# Core dependencies
aiohttp>=3.9.0
aiofiles>=23.2.1
backoff>=2.2.1
pydantic>=2.5.0
python-dotenv>=1.0.0
structlog>=24.1.0

# OpenAI API
openai>=1.6.0
httpx>=0.24.0  # Required for AsyncOpenAI timeout and connection pooling

# Optional for enhanced logging
colorama>=0.4.6  # For colored console output on Windows
rich>=13.7.0      # For better console formatting

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development tools (optional)
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.0
